Made LaunchScreen look better
Wasn't able to get the launch screen view controller working, so we will have to go with this. It looks like when you first download it to the phone, but after that it only shows the screen for a split second. I think if our app had more functionality, it would display the launch screen for a longer period of time